refactor: repository move FROM each sever TO common server
