+++ isol/server-child/src/main/java/shinhan/server_child/domain/user/controller/UserController.java
@@ -6,14 +6,14 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import shinhan.server_child.domain.user.dto.*;
 import shinhan.server_child.domain.user.service.UserService;
+import shinhan.server_common.domain.user.dto.*;
+import shinhan.server_common.global.exception.AuthException;
 import shinhan.server_common.global.security.JwtService;
 import shinhan.server_common.global.security.dto.FamilyInfoResponse;
 import shinhan.server_common.global.security.dto.JwtTokenResponse;
 import shinhan.server_common.global.security.dto.UserInfoResponse;
 import shinhan.server_common.global.utils.ApiUtils;
-import shinhan.server_common.global.exception.AuthException;
 
 import java.util.List;
 import java.util.Set;
