--- isol/server-child/src/main/java/shinhan/server_child/domain/user/dto/AccessTokenRequest.java
@@ -1,13 +1,0 @@
-package shinhan.server_child.domain.user.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class AccessTokenRequest {
-
-    @NotBlank(message = "refreshToken을 입력해주세요.")
-    private String refreshToken;
-}
