--- isol/server-child/src/main/java/shinhan/server_child/domain/user/dto/AccessTokenResponse.java
@@ -1,11 +1,0 @@
-package shinhan.server_child.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class AccessTokenResponse {
-
-    private String accessToken;
-}
