+++ isol/server-common/src/main/java/shinhan/server_common/domain/user/dto/ChildUpdateRequest.java
@@ -1,4 +1,4 @@
-package shinhan.server_parent.domain.user.dto;
+package shinhan.server_common.domain.user.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -13,22 +13,22 @@ import java.sql.Date;
 @NoArgsConstructor
 @Getter
 @Setter
-public class ParentsUpdateRequest {
+public class ChildUpdateRequest {
 
     private long serialNum;
     @NotBlank(message = "전화번호를 입력해주세요.")
     @Pattern(regexp = "\\d{3}-\\d{4}-\\d{4}", message = "전화번호 형식이 올바르지 않습니다.")
-    private String phoneNum;
+    private  String phoneNum;
     @NotBlank(message = "이름을 입력해주세요.")
     @Pattern(regexp = "^[가-힣]{2,5}$", message = "이름은 한글로 최소 2글자 최대 5글자까지 입력 가능합니다.")
-    private String name;
+    private  String name;
     @NotNull(message = "생일을 입력해주세요.")
     @Past(message = "생일은 현재 날짜보다 이전이어야 합니다.")
-    private Date birthDate;
+    private  Date birthDate;
     @NotNull(message = "프로필 번호를 입력해주세요.")
-    private int profileId;
+    private  int profileId;
 
-    public ParentsUpdateRequest(String phoneNum, String name, Date birthDate, int profileId) {
+    public ChildUpdateRequest(String phoneNum, String name, Date birthDate, int profileId) {
         this.phoneNum = phoneNum;
         this.name = name;
         this.birthDate = birthDate;
