+++ isol/server-common/src/main/java/shinhan/server_common/domain/user/dto/FamilyFindOneResponse.java
@@ -1,4 +1,4 @@
-package shinhan.server_parent.domain.user.dto;
+package shinhan.server_common.domain.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
