+++ isol/server-common/src/main/java/shinhan/server_common/domain/user/dto/FamilyInfoInterface.java
@@ -1,4 +1,4 @@
-package shinhan.server_child.domain.user.dto;
+package shinhan.server_common.domain.user.dto;
 
 public interface FamilyInfoInterface {
