+++ isol/server-common/src/main/java/shinhan/server_common/domain/user/repository/ParentsRepository.java
@@ -1,9 +1,9 @@
-package shinhan.server_parent.domain.user.repository;
+package shinhan.server_common.domain.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.jpa.repository.query.Procedure;
-import shinhan.server_parent.domain.user.entity.Parents;
+import shinhan.server_common.domain.user.entity.Parents;
 
 import java.util.List;
 import java.util.Optional;
