--- isol/server-parent/src/main/java/shinhan/server_parent/domain/user/dto/ChildFindOneResponse.java
@@ -1,36 +1,0 @@
-package shinhan.server_parent.domain.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.stereotype.Component;
-import shinhan.server_parent.domain.user.entity.Child;
-
-import java.sql.Date;
-
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Component
-public class ChildFindOneResponse {
-
-    private long serialNum;
-    private String phoneNum;
-    private String name;
-    private Date birthDate;
-    private int profileId;
-    private int score;
-
-    public static ChildFindOneResponse from(Child child) {
-        return ChildFindOneResponse.builder()
-                .serialNum(child.getSerialNum())
-                .phoneNum(child.getPhoneNum())
-                .name(child.getName())
-                .birthDate(child.getBirthDate())
-                .profileId(child.getProfileId())
-                .score(child.getScore())
-                .build();
-    }
-}
