--- isol/server-parent/src/main/java/shinhan/server_parent/domain/user/dto/FamilyInfoInterface.java
@@ -1,7 +1,0 @@
-package shinhan.server_parent.domain.user.dto;
-
-public interface FamilyInfoInterface {
-
-    long getSn();
-    String getName();
-}
