--- isol/server-parent/src/main/java/shinhan/server_parent/domain/user/dto/PhoneFindRequest.java
@@ -1,17 +1,0 @@
-package shinhan.server_parent.domain.user.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class PhoneFindRequest {
-
-    @NotBlank(message = "전화번호를 입력해주세요.")
-    @Pattern(regexp = "\\d{3}-\\d{4}-\\d{4}", message = "전화번호 형식이 올바르지 않습니다.")
-    private String phoneNum;
-}
