+++ isol/server-parent/src/main/java/shinhan/server_parent/domain/user/service/UserService.java
@@ -6,15 +6,15 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import shinhan.server_parent.domain.user.dto.*;
-import shinhan.server_parent.domain.user.entity.Child;
-import shinhan.server_parent.domain.user.entity.Family;
-import shinhan.server_parent.domain.user.entity.Parents;
-import shinhan.server_parent.domain.user.repository.ChildRepository;
-import shinhan.server_parent.domain.user.repository.FamilyRepository;
-import shinhan.server_parent.domain.user.repository.ParentsRepository;
-import shinhan.server_common.global.security.dto.FamilyInfoResponse;
+import shinhan.server_common.domain.user.dto.*;
+import shinhan.server_common.domain.user.entity.Child;
+import shinhan.server_common.domain.user.entity.Family;
+import shinhan.server_common.domain.user.entity.Parents;
+import shinhan.server_common.domain.user.repository.ChildRepository;
+import shinhan.server_common.domain.user.repository.FamilyRepository;
+import shinhan.server_common.domain.user.repository.ParentsRepository;
 import shinhan.server_common.global.exception.AuthException;
+import shinhan.server_common.global.security.dto.FamilyInfoResponse;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -29,6 +29,13 @@ public class UserService {
     private ParentsRepository parentsRepository;
     private ChildRepository childRepository;
     private FamilyRepository familyRepository;
+
+    private static boolean isUpdated(ParentsUpdateRequest parentsUpdateRequest, Parents updatedParents) {
+        return updatedParents.getPhoneNum().equals(parentsUpdateRequest.getPhoneNum())
+                && updatedParents.getName().equals(parentsUpdateRequest.getName())
+                && updatedParents.getBirthDate().equals(parentsUpdateRequest.getBirthDate())
+                && updatedParents.getProfileId() == parentsUpdateRequest.getProfileId();
+    }
 
     @Transactional
     public ParentsFindOneResponse getUser(long sn) {
@@ -55,13 +62,6 @@ public class UserService {
         } else {
             throw new InternalError("회원 정보 변경이 실패하였습니다.");
         }
-    }
-
-    private static boolean isUpdated(ParentsUpdateRequest parentsUpdateRequest, Parents updatedParents) {
-        return updatedParents.getPhoneNum().equals(parentsUpdateRequest.getPhoneNum())
-                && updatedParents.getName().equals(parentsUpdateRequest.getName())
-                && updatedParents.getBirthDate().equals(parentsUpdateRequest.getBirthDate())
-                && updatedParents.getProfileId() == parentsUpdateRequest.getProfileId();
     }
 
     @Transactional
@@ -95,7 +95,7 @@ public class UserService {
     public ParentsFindOneResponse join(JoinInfoSaveRequest joinInfoSaveRequest) {
         long serialNum = parentsRepository.generateSerialNum();
         log.info("Generated serial number={}", serialNum);
-        Parents parents = parentsRepository.save(joinInfoSaveRequest.convertToUser(serialNum, passwordEncoder));
+        Parents parents = parentsRepository.save(joinInfoSaveRequest.convertToParents(serialNum, passwordEncoder));
 
         return parents.convertToUserFindOneResponse();
     }
@@ -118,7 +118,7 @@ public class UserService {
 
     @Transactional()
     public List<FamilyInfoResponse> getFamilyInfo(long sn) {
-        return familyRepository.findMyFamilyInfo(sn)
+        return familyRepository.findChildInfo(sn)
                 .stream()
                 .map(myFamily -> new FamilyInfoResponse(myFamily.getSn(), myFamily.getName()))
                 .collect(Collectors.toList());
